refactor: migrate 3 monkey-patch tests to per-test-bundle (batch A)
Per Hoie's adoption push — replace the monkey-patch + afterAll-restore
pattern with leaf-pg mocking in the bundled framework. Same assertions
shifted from model-method args to SQL parameters; same coverage.

Tests migrated
- routes/users/delete-item → delete-item.test.bundle.ts
- routes/accounts/post-suggest-category → post-suggest-category.test.bundle.ts
- routes/accounts/holding-snapshot-sibling-routes (split) →
  delete-holding-snapshot.test.bundle.ts +
  get-holding-snapshots.test.bundle.ts

The last one had to split: the original `.test.ts` covered TWO routes
(delete + get) sharing one mock setup, but the per-test-bundle pattern
requires one `@bundles` source per file. Each route now gets its own
bundled test + bundle output.

Assertion shape changes (precision trade-off documented in the PR
discussion): tests that previously asserted on a model method's args
(e.g. `mockTxUpdate.mock.calls[0]`) now walk `mockQuery.mock.calls`
for the matching SQL statement and check its parameter array. Less
column-order-precise, but pins the same data invariants.

Verified locally
- `bun run test:bundled`     → 15 / 146 / ~540ms
- `bun run test:non-bundled` → 34 / 505 / ~4.2s
- `bun run test`             → both, 651 / 49 green
- typecheck + lint clean

Follow-ups
- Batch B (3 harder monkey-patch tests: post-api-keys, post-login,
  post-public-token) — post-public-token needs plaid-module mocking
- Path 3 (retire DI from compute-tools sources + tests, one PR each)
