+++ src/server/routes/users/delete-item.test.bundle.ts
@@ -1,41 +1,35 @@
-/**
- * Tests for DELETE /api/item (#393 — partial: delete-item.ts coverage).
- *
- * Scope: auth gate, missing-id validation, and the cross-user ownership
- * check via `searchItems`. The plaid / non-plaid dispatch on the
- * happy-path (calls into `plaid.deleteItem` + the items.ts `deleteItem`
- * that wraps `withTransaction` + 6 different tables) is left for a
- * follow-up that introduces dependency injection at the route layer.
- * Mocking the namespace + every table here would be fragile and would
- * leak across sibling test files.
- */
+// Per-test-bundle isolation — see scripts/test-bundled/.
+// @bundles src/server/routes/users/delete-item.ts
+import { describe, test, expect, mock, beforeEach } from "bun:test";
 
-import { describe, test, expect, mock, beforeEach, afterAll } from "bun:test";
+const mockQuery = mock(async (_sql: string, _values?: unknown[]) => ({
+  rows: [] as unknown[],
+  rowCount: 0 as number | null,
+}));
 
-import { itemsTable } from "server/lib/postgres/models";
-import { deleteItemRoute } from "./delete-item";
+class FakePool {
+  query = mockQuery;
+  end = async () => {};
+  connect = async () => ({ query: mockQuery, release: () => {} });
+}
 
-const originalQuery = itemsTable.query.bind(itemsTable);
+mock.module("pg", () => ({
+  Pool: FakePool,
+  types: { setTypeParser: () => {} },
+  default: { Pool: FakePool, types: { setTypeParser: () => {} } },
+}));
 
-const mockQuery = mock(async (_filters: Record<string, unknown>): Promise<unknown[]> => []);
-
-(itemsTable as unknown as { query: typeof mockQuery }).query = mockQuery;
-
-afterAll(() => {
-  (itemsTable as unknown as { query: typeof originalQuery }).query = originalQuery;
-});
+const { deleteItemRoute } = await import("./delete-item");
 
 beforeEach(() => {
   mockQuery.mockReset();
-  mockQuery.mockResolvedValue([]);
 });
 
 function makeReq(
   query: Record<string, unknown>,
   opts: { user?: { user_id: string; username: string } | null } = {},
 ): Parameters<typeof deleteItemRoute.execute>[0] {
-  const user =
-    opts.user === undefined ? { user_id: "u-1", username: "test" } : opts.user;
+  const user = opts.user === undefined ? { user_id: "u-1", username: "test" } : opts.user;
   return {
     method: "DELETE",
     path: "/item",
@@ -66,6 +60,29 @@ const fakeRes = () =>
     end() {},
   }) as unknown as Parameters<typeof deleteItemRoute.execute>[1];
 
+/**
+ * Build a raw items-table row that `ItemModel`'s mapper will accept.
+ * The model needs every column to be present (null is fine, undefined
+ * is not), so this returns the full row shape with sensible defaults.
+ */
+const makeItemRow = (overrides: Record<string, unknown> = {}) => ({
+  item_id: "i-A",
+  user_id: "u-1",
+  access_token: "no_access_token",
+  institution_id: null,
+  available_products: null,
+  cursor: null,
+  status: null,
+  provider: "manual",
+  last_sync_status: null,
+  last_sync_at: null,
+  last_sync_error: null,
+  raw: null,
+  updated: null,
+  is_deleted: false,
+  ...overrides,
+});
+
 describe("delete-item", () => {
   test("rejects unauthenticated requests", async () => {
     const result = await deleteItemRoute.execute(
@@ -85,9 +102,8 @@ describe("delete-item", () => {
   });
 
   test("cross-user attempt: item not in caller's items → 'not owned' failure", async () => {
-    // User u-1 asks to delete item i-belongs-to-other. `searchItems(user)`
-    // returns the caller's items only — emulate "caller owns nothing".
-    mockQuery.mockResolvedValueOnce([]);
+    // User u-1 asks to delete an item that returns no rows in their scope.
+    mockQuery.mockResolvedValueOnce({ rows: [], rowCount: 0 });
 
     const result = await deleteItemRoute.execute(
       makeReq({ id: "i-belongs-to-other" }),
@@ -97,20 +113,18 @@ describe("delete-item", () => {
     expect(result?.message).toMatch(/not owned by the request user/);
 
     expect(mockQuery).toHaveBeenCalledTimes(1);
-    const [filters] = mockQuery.mock.calls[0] as [Record<string, unknown>];
-    expect(filters.user_id).toBe("u-1");
+    // SELECT was scoped to the caller's user_id — pinned at the SQL layer
+    // by `WHERE user_id = $N`.
+    const [, values] = mockQuery.mock.calls[0] as [string, unknown[]];
+    expect(values).toContain("u-1");
   });
 
   test("item exists but for a different item_id → 'not owned' failure", async () => {
     // Caller owns i-A, but requests deletion of i-B.
-    const otherRow = {
-      toJSON: () => ({
-        item_id: "i-A",
-        provider: "manual", // ItemProvider.MANUAL = "manual"
-        access_token: "no_access_token",
-      }),
-    };
-    mockQuery.mockResolvedValueOnce([otherRow]);
+    mockQuery.mockResolvedValueOnce({
+      rows: [makeItemRow({ item_id: "i-A" })],
+      rowCount: 1,
+    });
 
     const result = await deleteItemRoute.execute(makeReq({ id: "i-B" }), fakeRes());
     expect(result?.status).toBe("failed");
